+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - main page now includes pagination (50 devices at a time)
 - Allow setting the device type as part of the powershell updater script, default is "computer". 
 - Main inventory listing page now remembers filtering between page refreshes
+- can upload a CSV file to bulk import devices instead of manually adding one at a time
 
 ### Changed